requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
python-dotenv==1.0.1
apscheduler==3.10.4
mercadopago==2.2.2
qrcode==8.0
Pillow==11.1.0
